Add RecipeCardFragment and fix bugs

+++ app/src/main/java/com/example/bakingapp/models/Ingredient.java
@@ -5,17 +5,17 @@ import com.google.gson.annotations.SerializedName;
 public class Ingredient {
 
     @SerializedName("quantity")
-    private int quantity;
+    private double quantity;
     @SerializedName("measure")
     private String measure;
     @SerializedName("ingredient")
     private String name;
 
-    public int getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
 
@@ -35,7 +35,7 @@ public class Ingredient {
         this.name = name;
     }
 
-    public Ingredient(int quantity, String measure, String name) {
+    public Ingredient(double quantity, String measure, String name) {
         this.quantity = quantity;
         this.measure = measure;
         this.name = name;
